Refactor: modifying suggestions from @hoh to get everything working again.
