+++ vm_supervisor/pool.py
@@ -6,8 +6,7 @@ from aleph_message.models import ProgramContent, ProgramMessage
 
 from .conf import settings
 from .models import VmHash, VmExecution
-from .network.firewall import Firewall
-from .network.interfaces import TapInterface
+from .network.ip import Network
 
 logger = logging.getLogger(__name__)
 
@@ -24,12 +23,12 @@ class VmPool:
     counter: int  # Used to provide distinct ids to network interfaces
     executions: Dict[VmHash, VmExecution]
     message_cache: Dict[str, ProgramMessage] = {}
-    firewall: Firewall
+    network: Network
 
     def __init__(self):
         self.counter = settings.START_ID_INDEX
         self.executions = {}
-        self.firewall = Firewall()
+        self.network = Network()
 
     async def create_a_vm(
         self, vm_hash: VmHash, program: ProgramContent, original: ProgramContent
@@ -40,7 +39,7 @@ class VmPool:
         await execution.prepare()
         vm_id = self.get_unique_vm_id()
 
-        tap_interface = TapInterface.from_vm_id(vm_id=vm_id, firewall=self.firewall)
+        tap_interface = await self.network.get_or_create_tap(vm_id)
         await execution.create(vm_id=vm_id, tap_interface=tap_interface)
         return execution
