+++ vm_supervisor/supervisor.py
@@ -11,7 +11,6 @@ from typing import Awaitable, Callable
 
 from aiohttp import web
 
-from vm_supervisor.network.ip import network_instance
 from . import metrics
 from .conf import settings
 from .resources import about_system_usage
@@ -87,13 +86,11 @@ def run():
 
     try:
         if settings.ALLOW_VM_NETWORKING:
-            network_instance.initialize(
+            pool.network.initialize(
                 vm_address_pool_range=settings.IPV4_ADDRESS_POOL,
                 vm_network_size=settings.IPV4_NETWORK_SIZE,
                 external_interface=settings.NETWORK_INTERFACE,
             )
-            network_instance.enable_ipv4_forwarding()
-            pool.firewall.initialize_nftables()
 
         if settings.WATCH_FOR_MESSAGES:
             app.on_startup.append(start_watch_for_messages_task)
@@ -103,5 +100,4 @@ def run():
         web.run_app(app, host=settings.SUPERVISOR_HOST, port=settings.SUPERVISOR_PORT)
     finally:
         if settings.ALLOW_VM_NETWORKING:
-            pool.firewall.teardown_nftables()
-            network_instance.reset_ipv4_forwarding_state()
+            pool.network.teardown()
